pyproject.toml
[tool.poetry]
name = "audiotopia"
version = "0.1.0"
description = ""
authors = ["gaoxiaodong <zzrzhezhiren@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
django = "4.2.1"
djangorestframework = "3.14.0"
django-softdelete = "0.10.5"
django-mptt = "0.14.0"
setuptools = ">=67.8.0"
wheel = ">=0.40.0"
django-contrib-comments = ">=2.2.0"
django-debug-toolbar = "^4.1.0"
mysqlclient = "^2.1.1"
rest-framework-generic-relations = "^2.1.0"
django-filter = "^23.2"

pyannote-audio = { path = "./static/pyannote-audio-develop" }
urllib3 = "1.26.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "tsinghua-source"
# 清华源
url = "https://pypi.tuna.tsinghua.edu.cn/simple/"
#name = "aliyun-source"
#url = " http://mirrors.aliyun.com/pypi/simple/"
# 仅从定义仓库获取
default = false
